+++ jina/types/sets/__init__.py
@@ -1,5 +1,5 @@
 __copyright__ = "Copyright (c) 2020 Jina AI Limited. All rights reserved."
 __license__ = "Apache-2.0"
 
-from .document_set import DocumentSet
-from .querylang_set import QueryLangSet
+from .document import DocumentSet
+from .querylang import QueryLangSet
